tests/pixbuf-lowmem.c tests/pixbuf-random.c
2008-05-25  Björn Lindqvist  <bjourne@gmail.com>

	* tests/pixbuf-lowmem.c
	* tests/pixbuf-random.c
	* tests/pixbuf-randomly-modified.c
	* tests/pixbuf-read.c
	* tests/pixbuf-threads.c
	* tests/testgtk.c
	* tests/testtreeview.c: Fix remaining compiler warnings in these
	files (#523751).

svn path=/trunk/; revision=20144

+++ tests/testtreeview.c
@@ -1050,7 +1050,7 @@ gtk_real_model_types_get_value (GtkTreeModel *tree_model,
         
         g_value_init (value, G_TYPE_STRING);
 
-        str = g_strdup_printf ("%ld", type);
+        str = g_strdup_printf ("%ld", (long int) type);
         g_value_set_string (value, str);
         g_free (str);
       }
@@ -1249,7 +1249,9 @@ gtk_real_model_types_iter_parent (GtkTreeModel *tree_model,
   if (parent == G_TYPE_INVALID)
     {
       if (type > G_TYPE_FUNDAMENTAL_MAX)
-        g_warning ("no parent for %ld %s\n", type, g_type_name (type));
+        g_warning ("no parent for %ld %s\n",
+                   (long int) type,
+                   g_type_name (type));
       return FALSE;
     }
   else
